refactor: layout 컴포넌트를 다시 인라인으로
