--- src/components/common/Layout.tsx
@@ -1,14 +1,0 @@
-import { PropsWithChildren } from 'react';
-import { css, Theme } from '@emotion/react';
-
-export default function Layout({ children }: PropsWithChildren<{}>) {
-  return <div css={layoutCss}>{children}</div>;
-}
-
-const layoutCss = (theme: Theme) => css`
-  background: ${theme.color.background};
-  max-width: ${theme.size.maxWidth};
-  width: 100%;
-  margin: 0 auto;
-  padding: ${theme.size.layoutPadding};
-`;
